Scanner: independent FFT size and spectrum size
Forcing the two to be equal didn't really make sense. With this
independence, we can have a bunch of small FFTs of at low sample rates
combined into a bigger spectrum. Alternatively, we might want an FFT
size bigger than the spectrum size when the sample rate is high, just
so the capture duration at each frequency is reasonably long.